Move client_singleton.go to correct directory and fix incorrect messages
