+++ backend/main.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"fmt"
 	sw "github.com/ZPI-2024-25/KubernetesAccessManager/api"
-	"github.com/ZPI-2024-25/KubernetesAccessManager/common"
+	"github.com/ZPI-2024-25/KubernetesAccessManager/cluster"
 	"log"
 	"net/http"
 )
@@ -12,13 +12,14 @@ import (
 func main() {
 	flag.Parse()
 
-	_, err := common.GetInstance()
+	singleton, err := cluster.GetInstance()
 	if err != nil {
 		fmt.Printf("Error when loading config: %v\n", err)
 		return
 	}
 
 	log.Printf("Server started")
+	log.Printf("Authentication method: %s", singleton.GetAuthenticationMethod())
 	router := sw.NewRouter()
 	log.Fatal(http.ListenAndServe(":8080", router))
 }
